Core: moved tasks to task_workers.
